OxWOD (Alexa Skills - Lambda and Backend)
It is the full code to make OxWOD, Alexa Skills, work.

OxWodLambda - It is the code that lives on the AWS Lambda which is
triggered by user. It gets the current WOD info from DynamoDB.

OxWodWebDriver - It is Java desktop project that crawl the webpage and
get the WOD data. Then, it processes the data and uploads to DynamoDB.